Fix clang format file
IDEs often only check parent directories for a
project specific style file. So the file was
moved. Also, different versions of clang-format
expect different options for Standard. I removed
the options and we now rely on the options set
in the default google style, which is Cpp17 in
newer clang-format versions anyways. #20