Update GitHub workflow to use docker container
The CI/CD pipeline now uses a docker container to be executed on
GitHub-hosted hardware. It relies on a pre-built docker container living
on the GitHub Container Registry.

So far, the pre-built container has been created and deployed manually.
This could be improved in the future, but is good for now to get a
testing pipeline up and running as soon as possible.